+++ shell/shell.go
@@ -21,15 +21,26 @@ var inputs = map[string]InputFunc{
 	"set":        inSetValue,
 	"properties": inProperties,
 	"dump":       dumpData,
+	"level":      setLevel,
 }
 
+func setLevel(w *api.Wattpilot, data []string) {
+	if len(data) == 0 {
+		return
+	}
+	if err := w.ParseLogLevel(data[0]); err != nil {
+		fmt.Println("error on update: ", err)
+	}
+}
 func inStatus(w *api.Wattpilot, data []string) {
 	w.StatusInfo()
-
 	fmt.Println("")
 }
 
 func inGetValue(w *api.Wattpilot, data []string) {
+	if len(data) == 0 {
+		return
+	}
 	v, err := w.GetProperty(data[0])
 	if err != nil {
 		fmt.Println(err)
@@ -38,6 +49,9 @@ func inGetValue(w *api.Wattpilot, data []string) {
 	fmt.Println(v)
 }
 func inSetValue(w *api.Wattpilot, data []string) {
+	if len(data) <= 1 {
+		return
+	}
 	err := w.SetProperty(data[0], data[1])
 	if err == nil {
 		return
@@ -108,31 +122,47 @@ func dumpData(w *api.Wattpilot, data []string) {
 }
 
 func inConnect(w *api.Wattpilot, data []string) {
-	connected, err := w.Connect()
+	err := w.Connect()
 	if err != nil {
 		log.Println("Could not connect", err)
-	}
-	if !connected || !w.IsInitialized() {
-		return
 	}
 	log.Printf("Connected to WattPilot %s, Serial %s", w.GetName(), w.GetSerial())
 }
 
+func processUpdates(ups <-chan interface{}) {
+	updates = ups
+}
+
 var interrupt chan os.Signal
+
+var updates <-chan interface{}
 
 func main() {
 	host := os.Getenv("WATTPILOT_HOST")
 	pwd := os.Getenv("WATTPILOT_PASSWORD")
+	level := os.Getenv("WATTPILOT_LOG")
 	if host == "" || pwd == "" {
 		return
 	}
+	if level == "" {
+		level = "WARN"
+	}
 	w := api.New(host, pwd)
+	if err := w.ParseLogLevel(level); err != nil {
+		fmt.Println("Could not update loglevel to", level, err)
+	}
+	// just a sample to test notification updates
+	processUpdates(w.GetNotifications("fhz"))
 	inConnect(w, nil)
 
 	w.StatusInfo()
 
 	for {
 		select {
+
+		case <-updates:
+			fmt.Println(<-updates)
+			break
 
 		case <-interrupt:
 			w.Disconnect()